feat(tidy-listings): updated index to match mockups added scss stylelint updated lighthouse urls
