+++ .github/workflows/pr-s3.yml
@@ -194,7 +194,7 @@ jobs:
         with:
           urls: |
             https://panor.am/pr/${{github.event.number}}/
-            https://panor.am/pr/${{github.event.number}}/places/
+            https://panor.am/pr/${{github.event.number}}/places/don-peyo
           uploadArtifacts: false # save results as an action artifacts
           runs: 1
 
@@ -208,7 +208,6 @@ jobs:
         with:
           urls: |
             https://panor.am/pr/${{github.event.number}}/
-            https://panor.am/pr/${{github.event.number}}/places/
             https://panor.am/pr/${{github.event.number}}/places/don-peyo
           budgetPath: ./budget.json # test performance budgets
           uploadArtifacts: true # save results as an action artifacts
