--- content/places/don-peyo.json
@@ -1,17 +1,0 @@
-{
-  "name": "Don Peyo",
-  "slug": "don-peyo",
-  "subtitle": "Comida chilena típica",
-  "image": "don-peyo",
-  "location": {
-    "fullAddress": "Lo Encalada 465, Ñuñoa, Región Metropolitana, Chile",
-    "street": "Lo Encalada 465",
-    "commune": "Ñuñoa",
-    "region": "Región Metropolitana",
-    "coordinates": {
-      "latitude": -33.4566567,
-      "longitude": -70.6261882
-    }
-  },
-  "categories": ["chilena"]
-}
