+++ content/places/don-peyo.yaml
@@ -1,0 +1,14 @@
+name: Don Peyo
+slug: don-peyo
+subtitle: Comida chilena típica
+image: don-peyo
+location:
+  street: Lo Encalada 465
+  commune: Ñuñoa
+  region: Región Metropolitana
+  country: Chile
+  coordinates:
+    latitude: -33.4566567
+    longitude: -70.6261882
+categories:
+  - chilena
