--- content/places/el-buen-paladar-norteno.json
@@ -1,17 +1,0 @@
-{
-  "name": "El Buen Paladar Norteño",
-  "subtitle": "Comida Peruana",
-  "slug": "el-buen-paladar-norteno",
-  "image": "el-buen-paladar-norteno",
-  "location": {
-    "fullAddress": "Victoria 721, Santiago, Región Metropolitana, Chile",
-    "street": "Victoria 721",
-    "commune": "Santiago",
-    "region": "Región Metropolitana",
-    "coordinates": {
-      "latitude": -33.463119,
-      "longitude": -70.6430681
-    }
-  },
-  "categories": ["peruana"]
-}
