+++ content/places/el-buen-paladar-norteno.yaml
@@ -1,0 +1,14 @@
+name: El Buen Paladar Norteño
+slug: el-buen-paladar-norteno
+subtitle: Comida Peruana
+image: el-buen-paladar-norteno
+location:
+  street: Victoria 721
+  commune: Santiago
+  region: Región Metropolitana
+  country: Chile
+  coordinates:
+    latitude: -33.463119
+    longitude: -70.6430681
+categories:
+  - peruana
