--- content/places/pizzeria-la-argentina.json
@@ -1,17 +1,0 @@
-{
-  "name": "Pizzeria La Argentina",
-  "slug": "pizzeria-la-argentina",
-  "subtitle": "Mejor pizza al corte",
-  "image": "pizzeria-la-argentina",
-  "location": {
-    "fullAddress": "Av. Italia 989, Providencia, Región Metropolitana, Chile",
-    "street": "Av. Italia 989",
-    "commune": "Providencia",
-    "region": "Región Metropolitana",
-    "coordinates": {
-      "latitude": -33.4439959,
-      "longitude": -70.6279339
-    }
-  },
-  "categories": ["pizza"]
-}
