+++ package.json
@@ -68,7 +68,9 @@
     "sharp": "^0.27.1",
     "stylelint": "^13.10.0",
     "stylelint-config-prettier": "^8.0.2",
+    "stylelint-config-recommended-scss": "^4.2.0",
     "stylelint-config-standard": "^20.0.0",
+    "stylelint-scss": "^3.19.0",
     "ts-jest": "^26.5.1",
     "vue-jest": "^3.0.4"
   }
