+++ stylelint.config.js
@@ -1,9 +1,5 @@
 module.exports = {
-  extends: [
-    'stylelint-config-standard',
-    // 'stylelint-config-recommended-scss',
-    'stylelint-config-prettier',
-  ],
+  extends: ['stylelint-config-recommended-scss'],
   // add your custom config here
   // https://stylelint.io/user-guide/configuration
   rules: {},
